refactor(client): auth routes signin, signup
